Upgrade jest. Sync examples.
