--- react-native-fast-image-example-cocoapods/FastImage/DefaultImageGrid.js
@@ -1,16 +1,0 @@
-// @flow
-import React from 'react'
-import { Image } from 'react-native'
-import Icon from './Icons/Icon'
-import ImageGrid from './ImageGrid'
-
-const DefaultImageGrid = () => <ImageGrid ImageComponent={Image} />
-
-DefaultImageGrid.navigationOptions = {
-  tabBarLabel: 'Image Grid',
-  tabBarIcon: props => (
-    <Icon name="ios-image-outline" focusedName="ios-image" {...props} />
-  ),
-}
-
-export default DefaultImageGrid
