+++ react-native-fast-image-example-cocoapods/src/Button.js
@@ -1,0 +1,28 @@
+import React from 'react'
+import { StyleSheet, Text, TouchableOpacity, View } from 'react-native'
+
+const Button = ({ text, onPress }) => (
+    <TouchableOpacity onPress={onPress}>
+        <View style={styles.button}>
+            <Text style={styles.text}>{text}</Text>
+        </View>
+    </TouchableOpacity>
+)
+
+const styles = StyleSheet.create({
+    button: {
+        backgroundColor: 'black',
+        margin: 10,
+        height: 44,
+        paddingLeft: 10,
+        paddingRight: 10,
+        borderRadius: 10,
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    text: {
+        color: 'white',
+    },
+})
+
+export default Button
