+++ react-native-fast-image-example-cocoapods/src/FastImageGrid.js
@@ -1,0 +1,15 @@
+import React from 'react'
+import FastImage from 'react-native-fast-image'
+import Icon from './Icons/Icon'
+import ImageGrid from './ImageGrid'
+
+const FastImageGrid = () => <ImageGrid ImageComponent={FastImage} />
+
+FastImageGrid.navigationOptions = {
+    tabBarLabel: 'FastImage Grid',
+    tabBarIcon: props => (
+        <Icon name="ios-photos-outline" focusedName="ios-photos" {...props} />
+    ),
+}
+
+export default FastImageGrid
