+++ react-native-fast-image-example/src/Icons/Icon.js
